refactor: reorganize type imports and move types to dedicated files
