+++ components/about/timelines/experience-timeline.tsx
@@ -4,7 +4,7 @@ import { motion } from "framer-motion";
 import { SectionHeader } from "@/components/about/section-header";
 import { TimelineItem } from "@/components/about/timelines/timeline-item";
 import { containerVariants, itemVariants } from "@/components/about/variants";
-import { ExperienceItems } from "@/components/about/types";
+import { ExperienceItems } from "@/types/about";
 
 interface ExperienceTimelineProps {
   experience: ExperienceItems;
