+++ components/contact/contact-form/index.tsx
@@ -5,7 +5,7 @@ import { Input, Textarea, Button } from "@heroui/react";
 import { Icon } from "@iconify/react";
 import { motion, AnimatePresence } from "framer-motion";
 
-import { ContactFormField, ContactFormProps } from "@/components/contact/types";
+import { ContactFormField, ContactFormProps } from "@/types/contact";
 import { useContactForm } from "@/hooks/use-contact-form";
 import { SuccessMessage } from "@/components/contact/contact-form/success-message";
 
@@ -164,3 +164,4 @@ export const ContactForm: React.FC<ContactFormProps> = ({
     </AnimatePresence>
   );
 };
+
