+++ types/index.ts
@@ -1,0 +1,4 @@
+export * from "./about";
+export * from "./contact";
+export * from "./projects";
+export * from "./testimonials";
