+++ types/testimonials.ts
@@ -1,0 +1,9 @@
+export interface Testimonial {
+  readonly id: number;
+  readonly name: string;
+  readonly role: string;
+  readonly content: string;
+  readonly avatar: string;
+}
+
+export type TestimonialCardProps = Omit<Testimonial, "id">;
